+++ sdk-py/OuterRail/models/runnable_distribution.py
@@ -7,7 +7,8 @@ from typing import Any, ClassVar, Dict, List, Optional, Set
 from pydantic import BaseModel, ConfigDict, Field, StrictInt
 
 
-class Distribution(BaseModel):
+
+class RunnableDistribution(BaseModel):
 
     platform: Optional[Annotated[str, Field(strict=True, max_length=64)]] = Field(
         default=None,
@@ -27,7 +28,7 @@ class Distribution(BaseModel):
         Annotated[str, Field(strict=True, max_length=64)]
     ] = Field(default=None, description="The checksum algorithm.")
     copyright: Optional[Annotated[str, Field(strict=True, max_length=256)]] = None
-    content_url: Annotated[str | None, Field(strict=True, max_length=256)]
+    content_url: Optional[Annotated[str, Field(strict=True, max_length=256)]] = None
     content_size_kb: Optional[StrictInt] = None
     date_published: Optional[datetime] = Field(
         default=None,
@@ -42,6 +43,50 @@ class Distribution(BaseModel):
         default=None,
         description="The technology readiness level (TRL) of the distribution. TRL 1 is the lowest and stands for 'Basic principles observed', TRL 9 is the highest and stands for 'actual system proven in operational environment'.",
     )
+    installation_script: Optional[
+        Annotated[str, Field(strict=True, max_length=256)]
+    ] = Field(
+        default=None,
+        description="An url pointing to a script that can be run to setup the environment necessary for running this distribution. This can be a relative url, if this distribution is a file archive.",
+    )
+    installation: Optional[Annotated[str, Field(strict=True, max_length=1800)]] = Field(
+        default=None,
+        description="A human readable explanation of the installation, primarily meant as alternative for when there is no installation script.",
+    )
+    installation_time_milliseconds: Optional[StrictInt] = Field(
+        default=None,
+        description="An illustrative time that the installation might typically take.",
+    )
+    deployment_script: Optional[
+        Annotated[str, Field(strict=True, max_length=256)]
+    ] = Field(
+        default=None,
+        description="An url pointing to a script that can be run to use this resource. This can be a relative url, if this distribution is a file archive.",
+    )
+    deployment: Optional[Annotated[str, Field(strict=True, max_length=1800)]] = Field(
+        default=None,
+        description="A human readable explanation of the deployment, primarily meant as alternative for when there is no installation script.",
+    )
+    deployment_time_milliseconds: Optional[StrictInt] = Field(
+        default=None,
+        description="An illustrative time that the deployment might typically take.",
+    )
+    os_requirement: Optional[
+        Annotated[str, Field(strict=True, max_length=256)]
+    ] = Field(
+        default=None,
+        description="A human readable explanation for the required os.",
+    )
+    dependency: Optional[Annotated[str, Field(strict=True, max_length=256)]] = Field(
+        default=None,
+        description="A human readable explanation of (software) dependencies.",
+    )
+    hardware_requirement: Optional[
+        Annotated[str, Field(strict=True, max_length=256)]
+    ] = Field(
+        default=None,
+        description="A human readable explanation of hardware requirements.",
+    )
     __properties: ClassVar[List[str]] = [
         "platform",
         "platform_resource_identifier",
@@ -55,6 +100,15 @@ class Distribution(BaseModel):
         "encoding_format",
         "name",
         "technology_readiness_level",
+        "installation_script",
+        "installation",
+        "installation_time_milliseconds",
+        "deployment_script",
+        "deployment",
+        "deployment_time_milliseconds",
+        "os_requirement",
+        "dependency",
+        "hardware_requirement",
     ]
 
     model_config = ConfigDict(
@@ -74,7 +128,7 @@ class Distribution(BaseModel):
 
     @classmethod
     def from_json(cls, json_str: str) -> Optional[Self]:
-        """Create an instance of Distribution from a JSON string"""
+        """Create an instance of RunnableDistribution from a JSON string"""
         return cls.from_dict(json.loads(json_str))
 
     def to_dict(self) -> Dict[str, Any]:
@@ -98,7 +152,7 @@ class Distribution(BaseModel):
 
     @classmethod
     def from_dict(cls, obj: Optional[Dict[str, Any]]) -> Optional[Self]:
-        """Create an instance of Distribution from a dict"""
+        """Create an instance of RunnableDistribution from a dict"""
         if obj is None:
             return None
 
@@ -119,6 +173,17 @@ class Distribution(BaseModel):
                 "encoding_format": obj.get("encoding_format"),
                 "name": obj.get("name"),
                 "technology_readiness_level": obj.get("technology_readiness_level"),
+                "installation_script": obj.get("installation_script"),
+                "installation": obj.get("installation"),
+                "installation_time_milliseconds": obj.get(
+                    "installation_time_milliseconds"
+                ),
+                "deployment_script": obj.get("deployment_script"),
+                "deployment": obj.get("deployment"),
+                "deployment_time_milliseconds": obj.get("deployment_time_milliseconds"),
+                "os_requirement": obj.get("os_requirement"),
+                "dependency": obj.get("dependency"),
+                "hardware_requirement": obj.get("hardware_requirement"),
             }
         )
         return _obj